chore(agent): move agent and sensor to different packages
